+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/.env
@@ -1,0 +1,6 @@
+MONGO_USERNAME=root
+MONGO_PASSWORD=example
+MONGO_HOST=db
+MONGO_PORT=27017
+BG_COLOR=teal
+FAIL_FLAG=False
