+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/Dockerfile
@@ -6,6 +6,6 @@ COPY requirements.txt requirements.txt
 
 RUN pip3 install --no-cache-dir --requirement requirements.txt
 
-COPY . .
+COPY ./application .
 
 CMD ["python3", "-m" , "flask", "run", "--host=0.0.0.0"]
