+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/MongoData/.mongodb/mongosh/641c664a9012feb8e7171acc_log
@@ -1,0 +1,8 @@
+{"t":{"$date":"2023-03-23T14:46:34.416Z"},"s":"I","c":"MONGOSH","id":1000000000,"ctx":"log","msg":"Starting log","attr":{"execPath":"/usr/bin/mongosh","envInfo":{"EDITOR":null,"NODE_OPTIONS":null,"TERM":null},"version":"1.8.0","distributionKind":"compiled","buildArch":"x64","buildPlatform":"linux","buildTarget":"linux-x64","buildTime":"2023-02-28T13:57:25.355Z","gitVersion":"9cf53bc336c79e505cf034bf5e6f3b3b3796cf25","nodeVersion":"v16.19.1","opensslVersion":"1.1.1t+quic","sharedOpenssl":false}}
+{"t":{"$date":"2023-03-23T14:46:34.424Z"},"s":"I","c":"MONGOSH","id":1000000048,"ctx":"config","msg":"Loading global configuration file","attr":{"filename":"/etc/mongosh.conf","found":false}}
+{"t":{"$date":"2023-03-23T14:46:34.443Z"},"s":"I","c":"DEVTOOLS-CONNECT","id":1000000042,"ctx":"mongosh-connect","msg":"Initiating connection attempt","attr":{"uri":"mongodb://127.0.0.1:27017/admin?directConnection=true&serverSelectionTimeoutMS=2000&appName=mongosh+1.8.0","driver":{"name":"nodejs|mongosh","version":"5.1.0"},"devtoolsConnectVersion":"1.4.4","host":"127.0.0.1:27017"}}
+{"t":{"$date":"2023-03-23T14:46:34.451Z"},"s":"I","c":"DEVTOOLS-CONNECT","id":1000000035,"ctx":"mongosh-connect","msg":"Server heartbeat succeeded","attr":{"connectionId":"127.0.0.1:27017"}}
+{"t":{"$date":"2023-03-23T14:46:34.466Z"},"s":"I","c":"DEVTOOLS-CONNECT","id":1000000037,"ctx":"mongosh-connect","msg":"Connection attempt finished"}
+{"t":{"$date":"2023-03-23T14:46:34.485Z"},"s":"I","c":"MONGOSH","id":1000000004,"ctx":"connect","msg":"Connecting to server","attr":{"session_id":"641c664a9012feb8e7171acc","userId":null,"telemetryAnonymousId":"641c664a3798738f6f57b3f2","connectionUri":"mongodb://<ip address>:27017/admin?directConnection=true&serverSelectionTimeoutMS=2000&appName=mongosh+1.8.0","is_atlas":false,"is_localhost":true,"is_do":false,"server_version":"6.0.5","node_version":"v16.19.1","mongosh_version":"1.8.0","server_os":"linux","server_arch":"x86_64","is_enterprise":false,"auth_type":null,"is_data_federation":false,"dl_version":null,"atlas_version":null,"is_genuine":true,"non_genuine_server_name":"mongodb","fcv":"6.0","api_version":null,"api_strict":null,"api_deprecation_errors":null}}
+{"t":{"$date":"2023-03-23T14:46:34.512Z"},"s":"I","c":"MONGOSH","id":1000000010,"ctx":"shell-api","msg":"Initialized context","attr":{"method":"setCtx","arguments":{}}}
+{"t":{"$date":"2023-03-23T14:46:34.708Z"},"s":"I","c":"MONGOSH-SNIPPETS","id":1000000019,"ctx":"snippets","msg":"Loaded snippets","attr":{"installdir":"/data/db/.mongodb/mongosh/snippets"}}
