+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/MongoData/.mongodb/mongosh/config
@@ -1,0 +1,1 @@
+{"userId":"641c664a3798738f6f57b3f2","telemetryAnonymousId":"641c664a3798738f6f57b3f2","enableTelemetry":true}
