+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/MongoData/.mongodb/mongosh/snippets/package.json
@@ -1,0 +1,1 @@
+{}
