+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/MongoData/WiredTiger
@@ -1,0 +1,2 @@
+WiredTiger
+WiredTiger 10.0.2: (December 21, 2021)
