+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/MongoData/WiredTiger.lock
@@ -1,0 +1,1 @@
+WiredTiger lock file
