+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/application/static/styles.css
@@ -1,0 +1,36 @@
+.contact form {
+    margin-top: 0px;
+    padding: 0px 30px;
+}
+.contact h3 {
+    color: #488ace;
+    text-align: center;
+    font-weight: 600;
+    margin-top: 0px;
+    margin-bottom: 0px;
+}
+
+.contact .formelement {
+    margin: 0 0px 10px;
+    width: 350px;
+    text-align: center;
+    margin-left: auto;
+    margin-right: auto;
+}
+.contact label {
+    display: none;
+}
+.contact input {
+    width: 100%;
+    border: none;
+    border: 1px solid #dddddd78;
+    padding: 10px;
+    border-radius: 4px;
+}
+.contact textarea#message {
+    width: 100%;
+    height: 122px;
+    border: 1px solid #dddddd78;
+    border-radius: 4px;
+    padding: 10px;
+}
