+++ Kubernetes/NGINXIngress-FlaskApp-MongoDB/docker-compose.yaml
@@ -1,0 +1,39 @@
+version: '3.1'  
+
+networks:
+  app-tier:
+    driver: bridge
+    ipam:
+      config:
+        - subnet: 172.18.0.0/24
+          gateway: 172.18.0.1
+
+services:
+
+  db:
+    image: mongo
+    restart: always
+    ports:
+      - 27017:27017
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: "${MONGO_USERNAME}"
+      MONGO_INITDB_ROOT_PASSWORD: "${MONGO_PASSWORD}"
+    volumes:
+      - ./MongoData:/data/db
+    networks:
+      app-tier:
+        ipv4_address: 172.18.0.2
+
+  web:
+    build:
+      context: ./
+      dockerfile: Dockerfile  
+    restart: always
+    ports:
+      - 5000:5000
+    env_file: .env
+    networks:
+      app-tier:
+        ipv4_address: 172.18.0.3
+    depends_on:
+      - "db"
